+++ js-api/src/entities.ts
@@ -1110,16 +1110,22 @@ export class Package extends Entity {
   }
 
   /**
-   * Deprecated. Use getSettings instead.
-   *  Returns properties for a package.
-  */
+   * @deprecated The {@link getProperties} should not be used. Use {@link settings} instead
+   */
   getProperties(): Promise<any> {
     return this.getSettings();
   }
 
-  /** Returns settings for a package. */
+  /**
+   * @deprecated The {@link getSettings} should not be used. Use {@link settings} instead
+   */
   getSettings(): Promise<Map<string, any>> {
     return api.grok_Package_Get_Settings(this.name);
+  }
+
+  /** Returns settings for a package. */
+  get settings(): {[index: string]: any} {
+    return api.grok_Package_Get_Settings_Sync(this.name);
   }
 
   /** Updates settings for a package. */
@@ -1301,8 +1307,8 @@ export class Property {
   set nullable(s: boolean) { api.grok_Property_Set_Nullable(this.dart, s); }
 
   /** Default value */
-  get defaultValue(): any { return api.grok_Property_Get_DefaultValue(this.dart); }
-  set defaultValue(s: any) { api.grok_Property_Set_DefaultValue(this.dart, s); }
+  get defaultValue(): any { return toJs(api.grok_Property_Get_DefaultValue(this.dart)); }
+  set defaultValue(s: any) { api.grok_Property_Set_DefaultValue(this.dart, toDart(s)); }
 
   /** Property editor */
   get editor(): string { return api.grok_Property_Get(this.dart, 'editor'); }
@@ -1358,7 +1364,7 @@ export class Property {
                 getter: PropertyGetter,
                 setter: PropertySetter,
                 defaultValue: any = null): Property {
-    return new Property(api.grok_Property(name, type, getter, setter, defaultValue));
+    return new Property(api.grok_Property(name, type, getter, setter, toDart(defaultValue)));
   }
 
   /** Creates an integer property */