+++ packages/SequenceTranslator/src/package.ts
@@ -22,7 +22,6 @@ import {polyToolEnumerateHelmUI} from './polytool/pt-enumerate-seq-dialog';
 import {_setPeptideColumn} from './polytool/utils';
 import {PolyToolCsvLibHandler} from './polytool/csv-to-json-monomer-lib-converter';
 import {ITranslationHelper} from './types';
-import {addContextMenuUI} from './utils/context-menu';
 import {PolyToolConvertFuncEditor} from './polytool/pt-convert-editor';
 import {CyclizedNotationProvider} from './utils/cyclized';
 import {getSeqHelper} from '@datagrok-libraries/bio/src/utils/seq-helper';
@@ -50,7 +49,7 @@ async function initSequenceTranslatorInt(): Promise<void> {
 
 //name: Oligo Toolkit
 //meta.icon: img/icons/toolkit.png
-//meta.browsePath: Oligo
+//meta.browsePath: Peptides | Oligo Toolkit
 //tags: app
 //output: view v
 export async function oligoToolkitApp(): Promise<DG.ViewBase> {
@@ -65,7 +64,7 @@ export async function oligoToolkitApp(): Promise<DG.ViewBase> {
 
 //name: Oligo Translator
 //meta.icon: img/icons/translator.png
-//meta.browsePath: Oligo
+//meta.browsePath: Peptides | Oligo Toolkit
 //tags: app
 //output: view v
 export async function oligoTranslatorApp(): Promise<DG.ViewBase> {
@@ -75,7 +74,7 @@ export async function oligoTranslatorApp(): Promise<DG.ViewBase> {
 
 //name: Oligo Pattern
 //meta.icon: img/icons/pattern.png
-//meta.browsePath: Oligo
+//meta.browsePath: Peptides | Oligo Toolkit
 //tags: app
 //output: view v
 export async function oligoPatternApp(): Promise<DG.ViewBase> {
@@ -85,7 +84,7 @@ export async function oligoPatternApp(): Promise<DG.ViewBase> {
 
 //name: Oligo Structure
 //meta.icon: img/icons/structure.png
-//meta.browsePath: Oligo
+//meta.browsePath: Peptides | Oligo Toolkit
 //tags: app
 //output: view v
 export async function oligoStructureApp(): Promise<DG.ViewBase> {
@@ -210,7 +209,7 @@ export async function getPolyToolConvertEditor(call: DG.FuncCall): Promise<DG.Co
 export async function polyToolConvert2(table: DG.DataFrame,
   seqCol: DG.Column, generateHelm: boolean, chiralityEngine: boolean, rules: string[]
 ): Promise<DG.Column<string>> {
-  const ptConvertRes = await polyToolConvert(seqCol, generateHelm, chiralityEngine, rules);
+  const ptConvertRes = await polyToolConvert(seqCol, generateHelm, false, chiralityEngine, false, rules);
   return ptConvertRes[0];
 }
 
@@ -248,14 +247,6 @@ export async function createMonomerLibraryForPolyTool(file: DG.FileInfo) {
   DG.Utils.download(jsonFileName, jsonFileContent);
 }
 
-// -- Handle context menu --
-
-//name: addContextMenu
-//input: object event
-export function addContextMenu(event: DG.EventData): void {
-  addContextMenuUI(event);
-}
-
 // //name: PolyTool Converter
 // //meta.icon: img/icons/structure.png
 // //meta.browsePath: PolyTool
@@ -282,21 +273,35 @@ export function addContextMenu(event: DG.EventData): void {
 //   }
 // }
 
-//name: PolyTool Enumerator Helm
+//name: HELM Enumerator
 //meta.icon: img/icons/structure.png
-//meta.browsePath: PolyTool
+//meta.browsePath: Peptides | PolyTool
 //tags: app
 export async function ptEnumeratorHelmApp(): Promise<void> {
   await polyToolEnumerateHelmUI();
 }
 
-//name: PolyTool Enumerator Chem
+//name: Chem Enumerator
 //meta.icon: img/icons/structure.png
-//meta.browsePath: PolyTool
+//meta.browsePath: Peptides | PolyTool
 //tags: app
 export async function ptEnumeratorChemApp(): Promise<void> {
   polyToolEnumerateChemUI();
 }
+
+
+//name: Polytool Helm Enumerator dialog
+//input: object cell {nullable: true}
+export async function getPtHelmEnumeratorDialog(cell?: DG.Cell) {
+  return polyToolEnumerateHelmUI(cell);
+}
+
+//name: Polytool Chem Enumerator dialog
+//input: object cell {nullable: true}
+export async function getPtChemEnumeratorDialog(cell?: DG.Cell) {
+  return polyToolEnumerateChemUI(cell);
+}
+
 
 //name: applyNotationProviderForHarmonizedSequence
 //input: column col