+++ packages/SequenceTranslator/src/polytool/conversion/pt-conversion.ts
@@ -6,13 +6,15 @@ import {Chain} from './pt-chain';
 import {_package} from '../../package';
 
 /** The main PolyTool convert engine. Returns list of Helms. Covered with tests. */
-export function doPolyToolConvert(sequences: string[], rules: Rules, helmHelper: IHelmHelper): string[] {
+export function doPolyToolConvert(sequences: string[], rules: Rules, helmHelper: IHelmHelper): [string[], boolean[]] {
   const helms = new Array<string>(sequences.length);
+  const isLinear = new Array<boolean>(sequences.length);
   for (let i = 0; i < sequences.length; i++) {
     try {
       if (sequences[i] == null) { helms[i] = ''; } else {
         const chain = Chain.fromSeparator(sequences[i], helmHelper);
         chain.applyRules(rules);
+        isLinear[i] = chain.monomersUnderRules.length > 1 || chain.linkagesUnderRules.length > 0 ? false : true;
         helms[i] = chain.getHelm();
       }
     } catch (err: any) {
@@ -21,5 +23,5 @@ export function doPolyToolConvert(sequences: string[], rules: Rules, helmHelper:
       helms[i] = '';
     }
   }
-  return helms;
+  return [helms, isLinear];
 }