+++ packages/UsageAnalysis/src/test-grid-cell-handler.ts
@@ -1,6 +1,8 @@
 import { category } from "@datagrok-libraries/utils/src/test";
 import { delay, DockManager, GridCell } from "datagrok-api/dg";
 import * as DG from 'datagrok-api/dg';
+import * as ui from 'datagrok-api/ui';
+import * as grok from 'datagrok-api/grok';
 
 export class TestGridCellHandler extends DG.ObjectHandler {
     get type(): string {