+++ packages/SequenceTranslator/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Sequence Translator changelog
 
+## 1.6.4 (WIP)
+
+PolyTool: linearization of single sequence molecules
+
+## 1.6.3 (2024-11-22)
+
+Moved Context menu items to detector
+
 ## 1.6.2 (2024-11-17)
 
 ### New features